+++ bonus/srcs/menu/add_dec_menu3.c
@@ -6,7 +6,7 @@
 /*   By: fmauguin <fmauguin@student.42.fr >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/05 22:17:37 by fmauguin          #+#    #+#             */
-/*   Updated: 2022/09/10 02:25:31 by fmauguin         ###   ########.fr       */
+/*   Updated: 2022/09/15 18:10:37 by fmauguin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #include "menu.h"
 #include "utils.h"
 
-void	add_dec_float(float *value, int op, t_delim lim)
+void	add_dec_float(double *value, int op, t_delim lim)
 {
 	if (op == 1)
 	{
